workers/transcript-worker/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
redis==5.0.1
nats-py==2.6.0
boto3==1.34.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
httpx==0.25.2
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# NLP and audio processing
openai==1.3.7
whisper==1.1.10
torch==2.1.1
torchaudio==2.1.1
transformers==4.36.0
sentence-transformers==2.2.2
spacy==3.7.2
pyannote.audio==3.1.1
librosa==0.10.1
soundfile==0.12.1
pydub==0.25.1

# Utilities
python-dateutil==2.8.2
pytz==2023.3
uuid==1.30

